requirements.txt
aiofiles==24.1.0
fastapi==0.116.1
openai-agents==0.2.8
python-dotenv==1.1.1
uvicorn==0.35.0
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0
pandas
numpy
matplotlib
PyMuPDF
chromadb
black
flake8